refactor(core): three names that outlived their subject
No behaviour.

isReadOnlyToolCall and isAgentSpawnToolCall used to be asked about a call, in
the loop, just before dispatching it. They are asked about a tool now, once,
when the toolset is assembled — so they are isReadOnlyTool and spawnsAnAgent,
and the local they gate is `tagged`, not `named`, since what withCallID adds is
the call's ID and no name at all.

Both compaction paths refused to shorten a conversation below three messages,
each with its own bare 3. It is shortestCompactable, said once with the reason:
an opening turn and its reply cannot be made shorter by summarising them.

aiconv_test.go outlived aiconv.go by one commit. Its one test is about
ChatMessage round-tripping, so it lives in message_test.go with the rest.
