+++ internal/core/message_test.go
@@ -3,6 +3,8 @@ package core
 import (
 	"strings"
 	"testing"
+
+	"github.com/genai-io/sdk-go/pkg/ai"
 )
 
 func TestBuildCompactionTextStripsSystemReminders(t *testing.T) {
@@ -53,5 +55,51 @@ func TestBuildCompactionTextStripsReminderWithEmbeddedCloseTag(t *testing.T) {
 
 	if strings.TrimSpace(text) != "Please summarize this coding conversation:\n\nUser: fix it" {
 		t.Fatalf("BuildCompactionText() = %q, should drop the whole reminder despite embedded close tag", text)
+	}
+}
+
+// What this conversion owes the SDK is everything the model left behind: the
+// per-endpoint decision about which of it can go back moved to ai.Model, which
+// drops or trims it on the way out. So what is checked here is that nothing is
+// lost on the way in — the test that used to sit here, a table of protocols
+// and their answers, now lives once in the SDK instead of once per application.
+func TestAModelsOwnStateSurvivesTheConversion(t *testing.T) {
+	chat := ChatMessage{
+		Role: ChatAssistant, Content: "done",
+		Thinking: "weighing it", ThinkingSignature: "sig-1",
+		Reasoning: []ReasoningItem{{ID: "r1", EncryptedContent: "opaque"}},
+		ToolCalls: []ToolCall{{ID: "c1", Name: "Read", Input: "{}"}},
+	}
+
+	msg, ok := chat.ToMessage()
+	if !ok {
+		t.Fatal("an assistant row converted to nothing")
+	}
+	out := []Message{msg}
+	if len(out) != 1 {
+		t.Fatalf("converted to %d messages, want 1", len(out))
+	}
+
+	// Replay order: reasoning first — Anthropic rejects a thinking block that
+	// does not lead, and a Responses call whose reasoning item does not
+	// precede it — then the answer, then the calls.
+	var kinds []ai.BlockType
+	for _, b := range out[0].Content {
+		kinds = append(kinds, b.Type)
+	}
+	want := []ai.BlockType{ai.BlockThinking, ai.BlockReasoning, ai.BlockText, ai.BlockToolCall}
+	if len(kinds) != len(want) {
+		t.Fatalf("blocks = %v, want %v", kinds, want)
+	}
+	for i := range want {
+		if kinds[i] != want[i] {
+			t.Fatalf("blocks = %v, want %v", kinds, want)
+		}
+	}
+
+	// And the signature travels with the thinking it proves. Whether the
+	// endpoint being asked can take it is ai.Model's to answer, not this.
+	if got := out[0].Content[0].Signature; got != "sig-1" {
+		t.Errorf("signature = %q, want it carried through", got)
 	}
 }
